Refactor: 필요없는 모듈 정리 - utils 모듈, constants 모듈 삭제하고 적절한 모듈로 배치
