+++ juju-admin/build.gradle.kts
@@ -15,9 +15,6 @@ val jasyptVersion: String by project
 dependencies {
     implementation(project(":juju-core:application"))
 
-    implementation(project(":juju-support:utils"))
-    implementation(project(":juju-support:constants"))
-
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:$swaggerVersion")
