+++ juju-api/build.gradle.kts
@@ -14,9 +14,6 @@ val swaggerVersion: String by project
 dependencies {
     implementation(project(":juju-core:application"))
 
-    implementation(project(":juju-support:utils"))
-    implementation(project(":juju-support:constants"))
-
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:$swaggerVersion")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
