+++ juju-core/application/build.gradle.kts
@@ -3,10 +3,7 @@ dependencies {
     runtimeOnly(project(":juju-infrastructure"))
 
     implementation(project(":juju-support:aws"))
-    implementation(project(":juju-support:utils"))
-    implementation(project(":juju-support:constants"))
 
     implementation("org.springframework:spring-context")
     implementation("org.springframework:spring-tx")
-//    implementation("jakarta.transaction:jakarta.transaction-api")
 }
