+++ juju-core/domain/build.gradle.kts
@@ -1,7 +1,4 @@
 dependencies {
-    implementation(project(":juju-support:utils"))
-    implementation(project(":juju-support:constants"))
-
     implementation("org.springframework:spring-context")
     implementation("org.springframework:spring-tx")
 }
