+++ juju-core/domain/src/main/kotlin/com/juloungjuloung/juju/common/utils/EnumFinder.kt
@@ -1,4 +1,4 @@
-package com.juloungjuloung.juju
+package com.juloungjuloung.juju.common.utils
 
 inline infix fun <reified E : Enum<E>, V> ((E) -> V).findBy(value: V): E? {
     return enumValues<E>().firstOrNull { this(it) == value }
