+++ juju-core/domain/src/main/kotlin/com/juloungjuloung/juju/domain/member/Member.kt
@@ -1,11 +1,8 @@
 package com.juloungjuloung.juju.domain.member
 
-import com.juloungjuloung.juju.member.MemberRole
-
 data class Member(
     val loginId: String,
     val loginPassword: String,
     val nickname: String,
-    val profileUrl: String? = null,
-    val roles: List<MemberRole>
+    val profileUrl: String? = null
 )
