+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/member/MemberEntity.kt
@@ -1,7 +1,6 @@
 package com.juloungjuloung.juju.entity.member
 
 import com.juloungjuloung.juju.entity.BaseEntity
-import com.juloungjuloung.juju.member.MemberRole
 import jakarta.persistence.Entity
 
 @Entity
@@ -9,6 +8,5 @@ class MemberEntity(
     val loginId: String,
     val loginPassword: String,
     val nickname: String,
-    val profileUrl: String? = null,
-    val roles: List<MemberRole>
+    val profileUrl: String? = null
 ) : BaseEntity()
