+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/product/ProductColorEntity.kt
@@ -1,7 +1,7 @@
 package com.juloungjuloung.juju.entity.product
 
+import com.juloungjuloung.juju.common.constant.ProductColorEnum
 import com.juloungjuloung.juju.entity.BaseEntity
-import com.juloungjuloung.juju.product.ProductColorEnum
 import jakarta.persistence.Entity
 import jakarta.persistence.ManyToOne
