+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/product/ProductMaterialEntity.kt
@@ -1,7 +1,7 @@
 package com.juloungjuloung.juju.entity.product
 
+import com.juloungjuloung.juju.common.constant.ProductMaterialEnum
 import com.juloungjuloung.juju.entity.BaseEntity
-import com.juloungjuloung.juju.product.ProductMaterial
 import jakarta.persistence.Entity
 import jakarta.persistence.ManyToOne
 
@@ -9,6 +9,6 @@ import jakarta.persistence.ManyToOne
 class ProductMaterialEntity(
     @ManyToOne
     val product: ProductEntity,
-    val material: ProductMaterial,
+    val material: ProductMaterialEnum,
     val additionalPrice: Int = 0
 ) : BaseEntity()
