+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/product/impl/BraceletEntity.kt
@@ -1,8 +1,8 @@
 package com.juloungjuloung.juju.entity.product.impl
 
+import com.juloungjuloung.juju.common.constant.CATEGORY_BRACELET
 import com.juloungjuloung.juju.domain.product.impl.Bracelet
 import com.juloungjuloung.juju.entity.product.ProductEntity
-import com.juloungjuloung.juju.product.CATEGORY_BRACELET
 import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
