+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/product/impl/EarringEntity.kt
@@ -1,8 +1,8 @@
 package com.juloungjuloung.juju.entity.product.impl
 
+import com.juloungjuloung.juju.common.constant.CATEGORY_EARRING
 import com.juloungjuloung.juju.domain.product.impl.Earring
 import com.juloungjuloung.juju.entity.product.ProductEntity
-import com.juloungjuloung.juju.product.CATEGORY_EARRING
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
