+++ juju-infrastructure/src/main/kotlin/com/juloungjuloung/juju/entity/product/impl/NecklaceEntity.kt
@@ -1,8 +1,8 @@
 package com.juloungjuloung.juju.entity.product.impl
 
+import com.juloungjuloung.juju.common.constant.CATEGORY_NECKLACE
 import com.juloungjuloung.juju.domain.product.impl.Necklace
 import com.juloungjuloung.juju.entity.product.ProductEntity
-import com.juloungjuloung.juju.product.CATEGORY_NECKLACE
 import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
