--- juju-support/constants/src/main/kotlin/com/juloungjuloung/juju/member/MemberRole.kt
@@ -1,7 +1,0 @@
-package com.juloungjuloung.juju.member
-
-enum class MemberRole {
-    NORMAL,
-    ORGANIZER, // 전체 관리자
-    PRODUCT_ORGANIZER // 상품 관리자 (상품 등록한 판매자)
-}
