+++ settings.gradle.kts
@@ -9,8 +9,6 @@ include(
 
     ":juju-infrastructure",
 
-    ":juju-support:utils",
-    ":juju-support:constants",
     ":juju-support:aws"
 )
